requirements.txt
certifi==2025.11.12
regex==2025.11.3
fsspec==2025.10.0
jsonschema-specifications==2025.9.1
openvino-telemetry==2025.2.0
pytz==2025.2
tzdata==2025.2
openvino==2024.6.0
setuptools==65.5.0
flatbuffers==25.9.23
gevent==25.9.1
attrs==25.4.0
pip==25.3
packaging==25.0
coloredlogs==15.0.1
rich==14.2.0
pillow==11.3.0
humanfriendly==10.0
tenacity==9.1.2
natsort==8.4.0
click==8.3.1
duckduckgo-search==8.1.1
zope-interface==8.1.1
psutil==7.1.3
protobuf==6.33.1
multidict==6.7.0
zope-event==6.1
pyyaml==6.0.3
lxml==6.0.2
flask-cors==6.0.1
diskcache==5.6.3
sentence-transformers==5.1.2
tqdm==4.67.1
fonttools==4.60.1
transformers==4.57.3
jsonschema==4.25.1
typing-extensions==4.15.0
anyio==4.11.0
platformdirs==4.5.0
cma==4.4.0
about-time==4.2.1
async-timeout==4.0.3
markdown-it-py==4.0.0
marshmallow==3.26.1
filelock==3.20.0
aiohttp==3.13.2
orjson==3.11.4
idna==3.11
matplotlib==3.10.7
threadpoolctl==3.6.0
xxhash==3.6.0
pyreadline3==3.5.4
charset-normalizer==3.4.4
alive-progress==3.3.0
pyparsing==3.2.5
greenlet==3.2.4
jinja2==3.1.6
openpyxl==3.1.5
werkzeug==3.1.3
flask==3.1.2
networkx==3.1
pydot==3.0.4
markupsafe==3.0.3
langgraph-checkpoint==3.0.1
pypdf2==3.0.1
jsonpointer==3.0.0
git-filter-repo==2.47.0
pydantic-core==2.41.5
requests==2.32.5
pycparser==2.23
pygments==2.19.2
nncf==2.18.0
pydantic==2.12.4
pydantic-settings==2.12.0
psycopg2-binary==2.9.11
python-dateutil==2.9.0.post0
openai==2.8.1
aiohappyeyeballs==2.6.1
urllib3==2.5.0
torch==2.4.1
torchaudio==2.4.1
pandas==2.3.3
itsdangerous==2.2.0
sqlalchemy==2.0.44
wrapt==2.0.1
cffi==2.0.0
et-xmlfile==2.0.0
optimum==2.0.0
jsonpatch==1.33
numpy==1.26.4
onnxruntime==1.23.2
yarl==1.22.0
onnx==1.19.1
onnxruntime-openvino==1.19.0
six==1.17.0
scipy==1.15.3
sympy==1.14.0
faiss-cpu==1.13.0
ninja==1.13.0
accelerate==1.12.0
ormsgpack==1.12.0
blinker==1.9.0
distro==1.9.0
autograd==1.8.0
frozenlist==1.8.0
scikit-learn==1.7.2
shellingham==1.5.4
joblib==1.5.2
kiwisolver==1.4.9
aiosignal==1.4.0
contourpy==1.3.2
deprecated==1.3.1
exceptiongroup==1.3.1
sniffio==1.3.1
mpmath==1.3.0
python-dotenv==1.2.1
hf-xet==1.2.0
langchain==1.1.0
langchain-core==1.1.0
langchain-openai==1.1.0
mypy-extensions==1.1.0
httpcore==1.0.9
langgraph-prebuilt==1.0.5
langgraph==1.0.4
langchain-classic==1.0.0
langchain-text-splitters==1.0.0
requests-toolbelt==1.0.0
referencing==0.37.0
huggingface-hub==0.36.0
rpds-py==0.29.0
httpx==0.28.1
zstandard==0.25.0
tokenizers==0.22.1
typer-slim==0.20.0
torchvision==0.19.1
peft==0.18.0
h11==0.16.0
primp==0.15.0
cycler==0.12.1
jiter==0.12.0
tiktoken==0.12.0
gevent-websocket==0.10.1
tabulate==0.9.0
typing-inspect==0.9.0
graphemeu==0.7.2
defusedxml==0.7.1
annotated-types==0.7.0
safetensors==0.7.0
dataclasses-json==0.6.7
pymoo==0.6.1.6
ml-dtypes==0.5.4
langsmith==0.4.48
colorama==0.4.6
httpx-sse==0.4.3
typing-inspection==0.4.2
langchain-community==0.4.1
propcache==0.4.1
llama-cpp-python==0.3.16
langgraph-sdk==0.2.10
sentencepiece==0.2.1
moocore==0.1.10
mdurl==0.1.2
